requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.0
ipython==8.5.0
scikit-learn==1.3.2
pandas==2.2.2
matplotlib==3.9.0
openpyxl==3.1.3
pydantic==1.10.12
requests==2.31.0  